chore(call_server): move Cli into ConfigOptionsWithVersion parsing logic to lib
