+++ rust/services/call/server_lib/Cargo.toml
@@ -17,7 +17,9 @@ call_engine = { workspace = true }
 call_host = { workspace = true }
 chain = { workspace = true }
 chain_client = { workspace = true }
+clap = { workspace = true, features = ["derive", "string"] }
 common = { workspace = true }
+config = { workspace = true }
 dashmap = { workspace = true }
 derive-new = { workspace = true }
 derive_more = { workspace = true }
@@ -40,6 +42,7 @@ tokio = { workspace = true }
 toml = { workspace = true }
 tower-http = { workspace = true, features = ["cors", "sensitive-headers", "validate-request"] }
 tracing = { workspace = true }
+version = { workspace = true }
 
 [dev-dependencies]
 assert-json-diff = { workspace = true }
@@ -49,6 +52,7 @@ ethers = { workspace = true }
 guest_wrapper = { workspace = true }
 lazy_static = { workspace = true }
 provider = { workspace = true }
+regex = { workspace = true }
 reqwest = { workspace = true }
 serde_json = { workspace = true, features = ["preserve_order"] }
 tempfile = { workspace = true }
