+++ rust/services/call/server_lib/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod cli;
 pub mod config;
 pub mod gas_meter;
 pub mod jwt;
@@ -11,6 +12,7 @@ mod proof;
 mod proving;
 mod ser;
 
+pub use cli::Cli;
 pub use config::{Config, ConfigBuilder};
 pub use handlers::{v_call::types as v_call, v_get_proof_receipt::types as v_get_proof_receipt};
 pub use server::{serve, server};
